--- Libirary_/src/main/java/SearchEngine/Controller.java
@@ -1,75 +1,0 @@
-package SearchEngine;
-
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.ListView;
-import javafx.scene.control.TextField;
-
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.ResourceBundle;
-import java.util.stream.Collectors;
-
-public class Controller implements Initializable {
-
-    ArrayList<String> words = new ArrayList<>(
-            Arrays.asList("test", "dog","Human", "Days of our life", "The best day",
-                    "Friends", "Animal", "Human", "Humans", "Bear", "Life",
-                    "This is some text", "Words", "222", "Bird", "Dog", "A few words",
-                    "Subscribe!", "SoftwareEngineeringStudent", "You got this!!",
-                    "Super Human", "Super", "Like")
-    );
-
-    @FXML
-    private TextField searchBar;
-
-    @FXML
-    private ListView<String> listView;
-    @Override
-    public void initialize(URL arg0, ResourceBundle arg1) {
-
-        listView.getItems().addAll(words);
-        try {
-
-
-        listView.getSelectionModel().selectedItemProperty().addListener(new ChangeListener<String>() {
-
-            @Override
-            public void changed(ObservableValue<? extends String> arg0, String arg1, String arg2) {
-
-                String currentFood = listView.getSelectionModel().getSelectedItem();
-
-                searchBar.setText(currentFood);
-
-            }
-        });
-        }catch (NullPointerException ex){
-        System.out.println(ex);
-        }
-    }
-    @FXML
-    void search(ActionEvent event) {
-        listView.getItems().clear();
-        listView.getItems().addAll(searchList(searchBar.getText(),words));
-    }
-    @FXML
-    void selectWord(ActionEvent event){
-        searchBar.setText(listView.getTypeSelector());
-    }
-
-
-    private List<String> searchList(String searchWords, List<String> listOfStrings) {
-
-        List<String> searchWordsArray = Arrays.asList(searchWords.trim().split(" "));
-
-        return listOfStrings.stream().filter(input -> {
-            return searchWordsArray.stream().allMatch(word ->
-                    input.toLowerCase().contains(word.toLowerCase()));
-        }).collect(Collectors.toList());
-    }
-}
