+++ Libirary_/src/main/java/module-info.java
@@ -5,4 +5,6 @@ module com.example.libirary_ {
 
     opens com.example.libirary_ to javafx.fxml;
     exports com.example.libirary_;
+    opens SearchEngine to javafx.fxml;
+    exports SearchEngine;
 }
